Deler opp i mindre komponenter
