+++ src/frontend/Sider/Behandling/Kjøreliste/RegistrerKjørelisteManuelt/LagredeKjørelister/LagredeKjørelister.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { Heading, VStack } from '@navikt/ds-react';
 
-import { LagretKjøreliste } from './LagretKjøreliste';
+import { LagretKjørelisteKort } from './LagretKjørelisteKort';
 import { useRegistrerKjøreliste } from '../../../../../context/RegistrerKjørelisteContext/RegistrerKjørelisteContext';
 
 export const LagredeKjørelister: React.FC = () => {
@@ -16,7 +16,7 @@ export const LagredeKjørelister: React.FC = () => {
         <VStack gap="space-16">
             <Heading size="medium">Kjørelister registrert i denne behandlingen</Heading>
             {kjørelisterLagretIBehandling.map((reise) => (
-                <LagretKjøreliste key={reise.id} kjøreliste={reise} />
+                <LagretKjørelisteKort key={reise.id} kjøreliste={reise} />
             ))}
         </VStack>
     );
