+++ src/frontend/Sider/Behandling/Kjøreliste/RegistrerKjørelisteManuelt/LagredeKjørelister/LagretKjørelisteKort.tsx
@@ -1,0 +1,179 @@
+import React, { FC, useState } from 'react';
+
+import { Box, VStack } from '@navikt/ds-react';
+import { BorderNeutralSubtle } from '@navikt/ds-tokens/js';
+
+import { LagretKjRelisteKortFooter } from './LagretKjørelisteKortFooter';
+import { useRegistrerKjøreliste } from '../../../../../context/RegistrerKjørelisteContext/RegistrerKjørelisteContext';
+import { useSteg } from '../../../../../context/StegContext';
+import { Feilmelding } from '../../../../../komponenter/Feil/Feilmelding';
+import {
+    Feil,
+    feiletRessursTilFeilmelding,
+} from '../../../../../komponenter/Feil/feilmeldingUtils';
+import { Skillelinje } from '../../../../../komponenter/Skillelinje';
+import { RessursStatus } from '../../../../../typer/ressurs';
+import {
+    KjørelisteDag,
+    KjørelisteUke,
+    ManuellRegistreringUkeDto,
+    ManueltInnsendtKjørelisteUke,
+} from '../typer';
+import { LagretKjørelisteKortHeader } from './LagretKjørelisteKortHeader';
+import { LagretKjørelisteUkerTabell } from './LagretKjørelisteUkerTabell';
+import { SlettKjørelisteModal } from './SlettKjørelisteModal';
+
+export const LagretKjørelisteKort: FC<{
+    kjøreliste: ManueltInnsendtKjørelisteUke;
+}> = ({ kjøreliste }) => {
+    const { erStegRedigerbart } = useSteg();
+    const { oppdaterKjøreliste, tilgjengeligeReiser } = useRegistrerKjøreliste();
+
+    const [redigerer, settRedigerer] = useState(false);
+    const [begrunnelseInput, settBegrunnelseInput] = useState('');
+    const [ukerIRedigering, settUkerIRedigering] = useState<KjørelisteUke[]>([]);
+    const [dagerPerUke, settDagerPerUke] = useState<Record<string, KjørelisteDag[]>>({});
+    const [visLeggTilUkerPanel, settVisLeggTilUkerPanel] = useState(false);
+    const [visSletteModal, settVisSletteModal] = useState(false);
+    const [laster, settLaster] = useState(false);
+    const [feilmelding, settFeilmelding] = useState<Feil | undefined>(undefined);
+
+    const tilgjengeligeUker =
+        tilgjengeligeReiser
+            .find((reise) => reise.fom === kjøreliste.reiseFom)
+            ?.uker.filter((uke) => !ukerIRedigering.some((u) => u.fom === uke.fom)) ?? [];
+
+    const startRedigering = () => {
+        settBegrunnelseInput(kjøreliste.begrunnelse ?? '');
+        settUkerIRedigering([...kjøreliste.innsendteUker]);
+        settDagerPerUke(
+            Object.fromEntries(
+                kjøreliste.innsendteUker.map((uke) => [
+                    uke.fom,
+                    uke.dager.map((dag) => ({ ...dag })),
+                ])
+            )
+        );
+        settRedigerer(true);
+    };
+
+    const avbrytRedigering = () => {
+        settRedigerer(false);
+        settVisLeggTilUkerPanel(false);
+        settFeilmelding(undefined);
+    };
+
+    const oppdaterDagForUke = (ukeFom: string, oppdatertDag: KjørelisteDag) => {
+        settDagerPerUke((prev) => ({
+            ...prev,
+            [ukeFom]: prev[ukeFom].map((dag) =>
+                dag.dato === oppdatertDag.dato ? oppdatertDag : dag
+            ),
+        }));
+    };
+
+    const leggTilUker = (nyeUker: ManuellRegistreringUkeDto[]) => {
+        const nyeDager: Record<string, KjørelisteDag[]> = {};
+        const nyeKjørelisteUker: KjørelisteUke[] = nyeUker.map((uke) => {
+            const dager: KjørelisteDag[] = uke.dager.map((dato) => ({
+                dato,
+                harKjørt: false,
+                parkeringsutgift: undefined,
+            }));
+            nyeDager[uke.fom] = dager;
+            return { ukenummer: uke.ukenummer, fom: uke.fom, tom: uke.tom, dager };
+        });
+
+        settDagerPerUke((prev) => ({ ...prev, ...nyeDager }));
+        settUkerIRedigering((prev) =>
+            [...prev, ...nyeKjørelisteUker].sort((a, b) => a.fom.localeCompare(b.fom))
+        );
+    };
+
+    const lagre = async () => {
+        if (laster) return;
+        settLaster(true);
+        settFeilmelding(undefined);
+
+        const respons = await oppdaterKjøreliste(kjøreliste.id, {
+            begrunnelse: begrunnelseInput || undefined,
+            uker: ukerIRedigering.map((uke) => ({
+                fom: uke.fom,
+                dager: dagerPerUke[uke.fom] ?? uke.dager,
+            })),
+        });
+
+        if (respons.status === RessursStatus.SUKSESS) {
+            settRedigerer(false);
+            settVisLeggTilUkerPanel(false);
+        } else {
+            settFeilmelding(feiletRessursTilFeilmelding(respons));
+        }
+
+        settLaster(false);
+    };
+
+    const ukerForRendering = redigerer ? ukerIRedigering : kjøreliste.innsendteUker;
+
+    const åpneSletteModal = () => {
+        settVisSletteModal(true);
+    };
+
+    const lukkSletteModal = () => {
+        settVisSletteModal(false);
+    };
+
+    const visLeggTilUker = () => {
+        settVisLeggTilUkerPanel(true);
+    };
+
+    const lukkLeggTilUker = () => {
+        settVisLeggTilUkerPanel(false);
+    };
+
+    return (
+        <Box background="default" borderColor="neutral-subtle" borderRadius="12" borderWidth="1">
+            <VStack gap="space-8" paddingBlock="space-16 space-8" paddingInline="space-24">
+                <LagretKjørelisteKortHeader
+                    kjøreliste={kjøreliste}
+                    erStegRedigerbart={erStegRedigerbart}
+                    redigerer={redigerer}
+                    startRedigering={startRedigering}
+                    åpneSletteModal={åpneSletteModal}
+                    begrunnelseInput={begrunnelseInput}
+                    settBegrunnelseInput={settBegrunnelseInput}
+                />
+                <SlettKjørelisteModal
+                    kjørelisteId={kjøreliste.id}
+                    visModal={visSletteModal}
+                    lukkModal={lukkSletteModal}
+                />
+            </VStack>
+
+            <Skillelinje borderColor={BorderNeutralSubtle} />
+
+            <Box paddingBlock="space-8 space-16" paddingInline="space-16">
+                <VStack gap="space-8">
+                    <LagretKjørelisteUkerTabell
+                        redigerer={redigerer}
+                        ukerForRendering={ukerForRendering}
+                        dagerPerUke={dagerPerUke}
+                        oppdaterDagForUke={oppdaterDagForUke}
+                    />
+                    <LagretKjRelisteKortFooter
+                        redigerer={redigerer}
+                        tilgjengeligeUker={tilgjengeligeUker}
+                        visLeggTilUkerPanel={visLeggTilUkerPanel}
+                        leggTilUker={leggTilUker}
+                        visLeggTilUker={visLeggTilUker}
+                        lukkLeggTilUker={lukkLeggTilUker}
+                        lagre={lagre}
+                        avbrytRedigering={avbrytRedigering}
+                        laster={laster}
+                    />
+                    <Feilmelding feil={feilmelding} />
+                </VStack>
+            </Box>
+        </Box>
+    );
+};
